… daemon/graphdriver/copy/copy_test.go
… daemon/graphdriver/overlay/overlay.go
… daemon/graphdriver/vfs/copy_linux.go
… daemon/graphdriver/vfs/copy_unsupported.go
… daemon/graphdriver/vfs/driver.go
… layer/layer_test.go