**- What I did**
I made copies concurrent for VFS, and Overlay1. 

**- How I did it**
I started with adding a benchmark to the tests to give me a baseline. On my test system that gave me about ~1081927574 ns/op. I slowly improved the performance throughout the commits, and increased performance to about 552743671 ns/op in the final tests. I parallizes 

**- How to verify it**
There are attached benchmarks, and tests

**- Description for the changelog**
Make VFS Faster.