@sargun 2018-10-13
FWIW, benchmark results (of the included synthetic benchmark):

```
XFS:
After improvements:
BenchmarkCopyDir-64    	      30	 804674003 ns/op
Before improvements:
BenchmarkCopyDir-64    	      20	1477743492 ns/op


EXT4
After improvements:
BenchmarkCopyDir-64    	      50	 578406619 ns/op
Before improvements:
BenchmarkCopyDir-64    	      20	1206396460 ns/op
```

This is a pretty significant performance improvement for most folks who want to use VFS.