@sargun 2018-10-13
For context of why we're suddenly so interested in this: https://www.spinics.net/lists/linux-unionfs/msg05779.html -- overlay2 is a pretty big pain in production.