@sargun 2018-10-17
Some more performance numbers:

```
sargun@ubuntu:~/go/src/github.com/docker/docker$ time docker run ubuntu:latest echo


real	0m0.576s
user	0m0.056s
sys	0m0.017s
sargun@ubuntu:~/go/src/github.com/docker/docker$ docker info -f {{.Driver}}
vfs
sargun@ubuntu:~/go/src/github.com/docker/docker$ time docker run ubuntu:latest echo


real	0m0.906s
user	0m0.057s
sys	0m0.015s
sargun@ubuntu:~/go/src/github.com/docker/docker$ docker image inspect ubuntu --format {{.Size}}
84114788
sargun@ubuntu:~/go/src/github.com/docker/docker$ docker image inspect ubuntu |jq '.[0].RootFS.Layers | length'
5

```