@sargun 2018-10-17
So, just to get some benchmarks, (The ones with TMPDIR are on XFS, the other on EXT4):
```
TMPDIR=/mnt/bench ./copy.be29f743ef -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	   20000	    346803 ns/op
BenchmarkCopyDir-64                      	      10	 874673948 ns/op
PASS
./copy.be29f743ef -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	   20000	    269299 ns/op
BenchmarkCopyDir-64                      	      10	 702427282 ns/op
PASS


158a50c4ff (HEAD) copy: Add small optimization to remove unncessary chmod call
TMPDIR=/mnt/bench ./copy.158a50c4ff  -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	   20000	    440871 ns/op
BenchmarkCopyDir-64                      	      10	1014970563 ns/op
PASS
$ ./copy.158a50c4ff  -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	   30000	    321325 ns/op
BenchmarkCopyDir-64                      	      10	 831401364 ns/op
PASS


24050449be (HEAD) copy: Process directory metadata concurrently
$ TMPDIR=/mnt/bench ./copy.24050449be  -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	   20000	    467087 ns/op
BenchmarkCopyDir-64                      	      10	1025676831 ns/op
PASS
$ ./copy.24050449be  -test.run none  -test.bench=.* -test.benchtime 2s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	   10000	    230710 ns/op
BenchmarkCopyDir-64                      	     100	 440939286 ns/op
PASS


e6ffc8a646 (HEAD) copy: Do dumb multiplexing of copiers, so we have parallel file copy
$ TMPDIR=/mnt/bench ./copy.e6ffc8a646  -test.run none  -test.bench=.* -test.benchtime 2s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	     500	   5114880 ns/op
BenchmarkCopyDir-64                      	       2	1165530186 ns/op
PASS
$ ./copy.e6ffc8a646  -test.run none  -test.bench=.* -test.benchtime 2s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	    1000	   2886859 ns/op
BenchmarkCopyDir-64                      	       3	2650059962 ns/op
PASS

6e48e3fb21 (HEAD) copy: Copy file data in the background, while copying metadata
$ TMPDIR=/mnt/bench ./copy.6e48e3fb21  -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	   20000	    300600 ns/op
BenchmarkCopyDir-64                      	      10	1097935740 ns/op
PASS
$ ./copy.6e48e3fb21  -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	   30000	    219280 ns/op
BenchmarkCopyDir-64                      	      10	 813604712 ns/op


e29bba98e1 (HEAD) copy: Refactor, and separate out metadata copy
$ TMPDIR=/mnt/bench ./copy.e29bba98e1  -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	    2000	   2523843 ns/op
BenchmarkCopyDir-64                      	       5	1503540458 ns/op
PASS
$ ./copy.e29bba98e1  -test.run none  -test.bench=.* -test.benchtime 5s
goos: linux
goarch: amd64
pkg: github.com/docker/docker/daemon/graphdriver/copy
BenchmarkCopyDirFewerDirsMoreFiles-64    	    3000	   1697842 ns/op
BenchmarkCopyDir-64                      	       5	1298545445 ns/op
PASS
```