@dmcgowan 2018-10-17
LGTM

Thanks for keeping the interface the same, documenting the code, adding tests, and benchmarks