@sargun 2018-10-18
@dmcgowan Mind nudging janky? Also, I'll freeze the PR as it is now, unless you want me to add more docs. If so, want to move it to review?