@sargun 2018-10-18
I'm going to follow this up with the OpenAt patches.

It'll get rid of:
* Extra allocations due to filepath manipulation
* go calling epoll_ctl (this takes ~200ns per file)