@sargun 2018-10-25
Can someone with tiny physical hardware, or spinning disks test this?

On Thu, Oct 25, 2018 at 8:33 AM Akihiro Suda <notifications@github.com>
wrote:

> *@AkihiroSuda* approved this pull request.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/moby/moby/pull/38034#pullrequestreview-168446984>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAtyREb0Y0CBFfCXAigehdsNcO0CorRrks5uodncgaJpZM4XahY0>
> .
>