@cpuguy83 2018-10-25
I'll fire this up on since and dual core systems with spinning disks.