@cpuguy83 2018-10-25
10s bench time times out on dual core now, here is no explicit bench time set between the two commits:

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkCopyDirFewerDirsMoreFiles-2     2195113       308630        -85.94%
```

Single core:
```
benchmark                              old ns/op     new ns/op     delta
BenchmarkCopyDirFewerDirsMoreFiles     280491        340916        +21.54%
```